microservices_mvp/order_service/requirements.txt
fastapi
uvicorn[standard]
requests 
# requests is included for future inter-service communication,
# though not used for direct calls in this initial MVP.
python-jose[cryptography]
